ProductsPage refactored
