--- Frontend/src/components/breadcrumbs/index.js
@@ -1,1 +1,0 @@
-export * from './BreadcrumbPage'
