--- Frontend/src/components/cards/index.js
@@ -1,1 +1,0 @@
-export * from './ProductCard'
