--- Frontend/src/components/filter/index.js
@@ -1,1 +1,0 @@
-export * from './FilterProducts'
