--- Frontend/src/components/inputs/index.js
@@ -1,4 +1,0 @@
-export * from './MenuCustom'
-export * from './MultiSelectCustom'
-export * from './OrderBySelect'
-export * from './SliderCustom'
