--- Frontend/src/components/paginator/index.js
@@ -1,1 +1,0 @@
-export * from './PaginatorCustom'
