+++ Frontend/src/helpers/index.js
@@ -1,3 +1,1 @@
 export * from './getEnvVariables'
-export * from './sameStrings'
-export * from './urlFunctions'
