+++ Frontend/src/main.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import { BrowserRouter } from 'react-router-dom'
 import { ApolloProvider } from '@apollo/client'
-import { apolloClient } from './api/apolloClient'
+import { apolloClient } from '@api/apolloClient'
 
 import { App } from './App'
