+++ Frontend/src/pages/Products/components/ProductCard.jsx
@@ -1,10 +1,8 @@
 import { useEffect, useState } from 'react'
 import { HStack, Image, Text, VStack, useColorModeValue } from '@chakra-ui/react'
 import { CardLink } from '@components'
-import { OfferRating, StarRating } from './rating'
-import { OFFER_RATING, STAR_RATING } from '../assets/ratingVariant'
 
-export const SliderCard = ({ dataCard, variant }) => {
+export const ProductCard = ({ dataCard, ...rest }) => {
   const [hrefCard, setHrefCard] = useState('')
 
   useEffect(() => {
@@ -13,7 +11,7 @@ export const SliderCard = ({ dataCard, variant }) => {
   }, [dataCard])
 
   return (
-    <CardLink href={hrefCard} my={3} mx={2}>
+    <CardLink href={hrefCard} {...rest}>
       <Image
         h={200}
         objectFit='cover'
@@ -26,22 +24,21 @@ export const SliderCard = ({ dataCard, variant }) => {
         spacing={0}
         alignItems='flex-start' justifyContent='space-between'
       >
-        <>
-          {
-            (variant === OFFER_RATING) && <OfferRating value={dataCard?.discount} />
-          }
-          {
-            (variant === STAR_RATING) && <StarRating value={dataCard?.dataValue} />
-          }
-        </>
-        <Text
-          fontWeight='medium' fontSize={18} minH={54}
-          color={useColorModeValue('light.text.main', 'dark.text.main')}
-        >
-          {dataCard?.title}
-        </Text>
-        <HStack w='full' justifyContent='space-between'>
+        <VStack spacing={0} alignItems='flex-start' w='full'>
           <Text color={useColorModeValue('light.text.secondary', 'dark.text.secondary')}>{dataCard?.brand}</Text>
+          <Text
+            fontWeight='medium' fontSize={18}
+            color={useColorModeValue('light.text.main', 'dark.text.main')}
+            minH='54px'
+          >
+            {dataCard?.title}
+          </Text>
+        </VStack>
+        <HStack w='full' justifyContent='space-between' alignItems='flex-end'>
+          <VStack spacing={0} alignItems='flex-start'>
+            <Text color={useColorModeValue('light.text.secondary', 'dark.text.secondary')}>Graduación: {dataCard?.alcoholic_grade}°</Text>
+            <Text color={useColorModeValue('light.text.secondary', 'dark.text.secondary')}>Contenido: {dataCard?.content}cc</Text>
+          </VStack>
           <Text
             fontWeight='medium' fontSize={20}
             color={useColorModeValue('light.text.main', 'dark.text.main')}
