+++ Frontend/src/pages/Products/components/index.js
@@ -1,0 +1,3 @@
+export * from './FilterProducts'
+export * from './ProductCard'
+export * from './ProductList'
