+++ Frontend/src/pages/Products/components/inputs/index.js
@@ -1,0 +1,4 @@
+export * from './MultiSelect'
+export * from '../OrderBySelect'
+export * from '../Paginator'
+export * from './Slider'
