+++ Frontend/src/pages/Products/helpers/index.js
@@ -1,0 +1,1 @@
+export * from './sameStrings'
