+++ Frontend/src/pages/Products/helpers/sameStrings.js
