+++ Frontend/src/pages/index.js
@@ -1,2 +1,1 @@
 export * from './Error404Page'
-export * from './ProductsPage'
